# src/server/routes/v1/openapi.yaml
openapi: 3.0.3
info:
  title: MeshMonitor API
  description: |
    MeshMonitor REST API v1 provides programmatic access to mesh network data.

    ## Multi-source (breaking change — spec 2.0.0)
    MeshMonitor 4.0 runs one installation against many Meshtastic/MQTT/MeshCore
    sources. Per-resource endpoints are now canonical under
    `/api/v1/sources/{sourceId}/...`:

    | Resource | Canonical path |
    |---|---|
    | Nodes | `/api/v1/sources/{sourceId}/nodes` |
    | Messages | `/api/v1/sources/{sourceId}/messages` |
    | Channels | `/api/v1/sources/{sourceId}/channels` |
    | Telemetry | `/api/v1/sources/{sourceId}/telemetry` |
    | Traceroutes | `/api/v1/sources/{sourceId}/traceroutes` |
    | Packets | `/api/v1/sources/{sourceId}/packets` |
    | Network | `/api/v1/sources/{sourceId}/network` |
    | Position history | `/api/v1/sources/{sourceId}/nodes/{nodeId}/position-history` |
    | Status | `/api/v1/sources/{sourceId}/status` |

    The literal string `default` may be used for `{sourceId}` — it resolves to
    the first source the authenticated token has `read` permission on (or the
    first enabled source for admin tokens). 404 if none is accessible.

    `GET /api/v1/sources` lists the sources the token can read.

    `/api/v1/solar` and `/api/v1/channel-database` remain deployment-global
    (no source scope).

    Legacy root paths (`/api/v1/nodes?sourceId=...`, etc.) were removed in
...
    Use the per-source shape instead.

    ## Authentication
...

    ```
...

...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Token
      description: |
        API token authentication. Use format: `Bearer mm_v1_<token>`

        Generate tokens from your User Settings page.
...
  schemas:
    Error:
      type: object
      required:
        - success
        - error
        - message
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Unauthorized"
        message:
          type: string
          example: "API token required"
...
    MessageListResponse:
      type: object
      required:
        - success
        - count
        - data
      properties:
        success:
          type: boolean
          example: true
        count:
          type: integer
          description: Number of messages returned
          example: 50
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    MessageResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Message'
    SendMessageRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: |
            The message text to send. Messages longer than 200 bytes will be
            automatically split into multiple parts and queued for delivery.
          example: "Hello from the API!"
          minLength: 1
          maxLength: 1000
        channel:
          type: integer
          description: Channel number to send to (0-7). Required if toNodeId is not specified.
          minimum: 0
          maximum: 7
          example: 0
        toNodeId:
          type: string
          description: Node ID for direct message (e.g., "!a1b2c3d4"). Required if channel is not specified.
          pattern: '^![0-9a-fA-F]{1,8}$'
          example: "!a1b2c3d4"
        replyId:
          type: integer
          description: Request ID of the message being replied to (optional)
          example: 123456789
    SendMessageResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required:
            - messageId
            - requestId
            - deliveryState
            - messageCount
          properties:
            messageId:
              type: string
              description: Unique message ID for tracking (format nodeNum_requestId)
              example: "2715451348_123456789"
            requestId:
              type: integer
              description: Request ID for matching delivery acknowledgments
              example: 123456789
            deliveryState:
              type: string
              description: Current delivery state
              enum: [pending, delivered, confirmed, failed]
              example: "pending"
            messageCount:
              type: integer
              description: Number of messages sent (always 1 for short messages)
              example: 1
            text:
              type: string
              description: The message text that was sent
              example: "Hello from the API!"
            channel:
              type: integer
              description: Channel the message was sent to (-1 for direct messages)
              example: 0
            toNodeId:
              type: string
              description: Destination node ID or "broadcast" for channel messages
              example: "broadcast"
...
paths:
...
  /messages:
    get:
      tags:
        - Messages
      summary: Get messages
      description: Retrieve mesh network messages with optional filters
      parameters:
        - name: channel
          in: query
          description: Filter by channel number
          required: false
          schema:
            type: integer
        - name: fromNodeId
          in: query
          description: Filter by sender node ID
          required: false
          schema:
            type: string
        - name: toNodeId
          in: query
          description: Filter by recipient node ID
          required: false
          schema:
            type: string
        - name: since
          in: query
          description: Unix timestamp to filter messages after this time
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum number of records to return
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Messages
      summary: Send a message
      description: |
        Send a new message to a channel or directly to a node.

        **Permission Requirements:**
        - Channel messages require `channel_X:write` permission (where X is the channel number)
        - Direct messages require `messages:write` permission

        **Usage:**
        - Provide `channel` for broadcast messages to a channel
        - Provide `toNodeId` for direct messages to a specific node
        - Do not provide both `channel` and `toNodeId`

        **Multi-Message Breakup:**
        - Meshtastic has a 200-byte payload limit for text messages
        - Messages longer than 200 bytes are automatically split into multiple parts (max 3 parts, ~600 bytes)
        - Messages requiring more than 3 parts will be rejected with HTTP 413 (Payload Too Large)
        - Split messages are queued and delivered with 30-second intervals between parts
        - Returns HTTP 202 (Accepted) when messages are queued for delivery
        - The response includes `messageCount` and `queueIds` for tracking split messages
        - Only the first message part will have the `replyId` set (if provided)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              channelMessage:
                summary: Send to a channel
                value:
                  text: "Hello from the API!"
                  channel: 0
              directMessage:
                summary: Send direct message
                value:
                  text: "Private message via API"
                  toNodeId: "!a1b2c3d4"
              replyMessage:
                summary: Reply to a message
                value:
                  text: "This is a reply"
                  channel: 0
                  replyId: 123456789
              longMessage:
                summary: Long message (will be split)
                description: Messages exceeding 200 bytes are automatically split and queued
                value:
                  text: "This is a longer message that exceeds the Meshtastic payload limit of 200 bytes. When sent via the API, it will be automatically split into multiple parts and queued for delivery with 30-second intervals between each part. This helps ensure reliable delivery of longer content."
                  channel: 0
      responses:
        '201':
          description: Message sent successfully (short message, ≤200 bytes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '202':
          description: Message queued for delivery (long message split into multiple parts)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageQueuedResponse'
        '400':
          description: Bad request (invalid parameters)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingText:
                  summary: Missing text
                  value:
                    success: false
                    error: "Bad Request"
                    message: "Message text is required"
                bothChannelAndNode:
                  summary: Both channel and toNodeId provided
                  value:
                    success: false
                    error: "Bad Request"
                    message: "Provide either channel OR toNodeId, not both"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Forbidden"
                  message:
                    type: string
                    example: "Insufficient permissions"
                  required:
                    type: object
                    properties:
                      resource:
                        type: string
                        example: "channel_0"
                      action:
                        type: string
                        example: "write"
        '413':
          description: Payload too large - message would require more than 3 parts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error: "Payload Too Large"
                message: "Message too long. Would require 5 parts but maximum is 3 parts (~600 bytes)"
        '503':
          description: Service unavailable - not connected to Meshtastic node
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                success: false
                error: "Service Unavailable"
                message: "Not connected to Meshtastic node"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Get a specific message
      description: Retrieve a specific message by ID
      parameters:
        - name: messageId
          in: path
          description: Message ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...